Refactor codebase to Model/View design
