+++ src/main/java/mario/model/Key.java
@@ -1,4 +1,4 @@
-package mario;
+package mario.model;
 
 public class Key {
