+++ src/main/java/mario/model/map/MapLoader.java
@@ -1,4 +1,4 @@
-package mario;
+package mario.model.map;
 
 import java.io.BufferedReader;
 import java.io.File;
