+++ src/main/java/mario/model/map/tiles/DoorTile.java
@@ -1,4 +1,4 @@
-package mario.tiles;
+package mario.model.map.tiles;
 
 import nl.han.ica.oopg.objects.Sprite;
 import nl.han.ica.oopg.tile.Tile;
