+++ src/main/java/mario/model/map/tiles/FloorTile.java
@@ -1,12 +1,11 @@
-package mario.tiles;
+package mario.model.map.tiles;
 
 import nl.han.ica.oopg.objects.Sprite;
 import nl.han.ica.oopg.tile.Tile;
 
-public class LavaTile extends Tile {
+public class FloorTile extends Tile {
 
-    public LavaTile(Sprite sprite) {
+    public FloorTile(Sprite sprite) {
         super(sprite);
     }
-
 }
