+++ src/main/java/mario/model/map/tiles/KeyTile.java
@@ -1,11 +1,12 @@
-package mario.tiles;
+package mario.model.map.tiles;
 
 import nl.han.ica.oopg.objects.Sprite;
 import nl.han.ica.oopg.tile.Tile;
 
-public class FloorTile extends Tile {
+public class KeyTile extends Tile {
 
-    public FloorTile(Sprite sprite) {
+    public KeyTile(Sprite sprite) {
         super(sprite);
     }
+
 }
