+++ src/main/java/mario/view/end/EndView.java
@@ -1,0 +1,4 @@
+package mario.view.end;
+
+public class EndView {
+}
