+++ src/main/java/mario/view/game/sprites/KeySprite.java
@@ -1,7 +1,6 @@
-package mario.dashboard;
+package mario.view.game.sprites;
 
 import mario.MainApp;
-import nl.han.ica.oopg.dashboard.Dashboard;
 import nl.han.ica.oopg.objects.Sprite;
 import nl.han.ica.oopg.objects.SpriteObject;
