+++ src/main/java/mario/view/menu/MenuView.java
@@ -1,0 +1,4 @@
+package mario.view.menu;
+
+public class MenuView {
+}
